fix: android type mismatch error on insert

+++ android/src/main/java/com/rnsqlite/Database.kt
@@ -54,13 +54,14 @@ class Database(context: Context, name: String, version: Int) :
     }
   }
 
-  fun getLastInsertRowId(): Int? {
+  fun getLastInsertRowId(): Long? {
     var cursor: Cursor? = null
     try {
       cursor = db.rawQuery("SELECT last_insert_rowid()", Array<String?>(0){null})
       if (cursor.moveToFirst()) {
-        if (cursor.columnCount > 0 && cursor.getType(0) == FIELD_TYPE_INTEGER)
-        return cursor.getInt(0)
+        if (cursor.columnCount > 0 &&
+            cursor.getType(0) == FIELD_TYPE_INTEGER)
+          return cursor.getLong(0)
       }
 
       return 0
